+++ R/auth.R
@@ -50,7 +50,7 @@ login <- function(email = NULL,
         prompt = paste("[crunch]", getOption("prompt")),
         crunch.old.prompt = getOption("prompt")
     )
-    message("Logged into crunch.io as ", email)
+    message("Logged into crunch.io as ", login_info$email)
     ## Return a Session object
     invisible(session())
 }
@@ -128,7 +128,7 @@ crunchAuth <- function(email, password = NULL, ...) {
     if (is.null(password)) {
         if (is.interactive()) {
             prompt <- paste0("Crunch.io password for ", email, ": ")
-            if ("rstudioapi" %in% rownames(installed.packages()) &&
+            if (requireNamespace("rstudioapi", quietly = TRUE) &&
                 rstudioapi::hasFun("askForPassword")) {
                 password <- rstudioapi::askForPassword(prompt)
             } else {